Refactor package structure for password recovery and token management modules
