+++ src/main/java/info/mackiewicz/bankapp/presentation/auth/controller/PasswordResetWebController.java
@@ -2,7 +2,7 @@ package info.mackiewicz.bankapp.presentation.auth.controller;
 
 import info.mackiewicz.bankapp.presentation.auth.dto.PasswordResetDTO;
 import info.mackiewicz.bankapp.presentation.auth.dto.PasswordResetRequestDTO;
-import info.mackiewicz.bankapp.system.security.recovery.password.service.PasswordResetTokenService;
+import info.mackiewicz.bankapp.system.recovery.password.service.PasswordResetTokenService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
