+++ src/test/java/info/mackiewicz/bankapp/shared/core/ApiExceptionHandlerTest.java
@@ -8,10 +8,10 @@ import info.mackiewicz.bankapp.shared.web.dto.ValidationError;
 import info.mackiewicz.bankapp.shared.web.error.mapping.ApiExceptionToErrorMapper;
 import info.mackiewicz.bankapp.shared.web.error.validation.ValidationErrorProcessor;
 import info.mackiewicz.bankapp.shared.web.util.RequestUriHandler;
-import info.mackiewicz.bankapp.system.security.exception.ExpiredTokenException;
-import info.mackiewicz.bankapp.system.security.exception.TokenNotFoundException;
-import info.mackiewicz.bankapp.system.security.exception.TooManyPasswordResetAttemptsException;
-import info.mackiewicz.bankapp.system.security.exception.UsedTokenException;
+import info.mackiewicz.bankapp.system.recovery.password.exception.ExpiredTokenException;
+import info.mackiewicz.bankapp.system.recovery.password.exception.TokenNotFoundException;
+import info.mackiewicz.bankapp.system.recovery.password.exception.TooManyPasswordResetAttemptsException;
+import info.mackiewicz.bankapp.system.recovery.password.exception.UsedTokenException;
 import info.mackiewicz.bankapp.user.exception.UserNotFoundException;
 import jakarta.validation.ConstraintViolationException;
 import org.junit.jupiter.api.BeforeEach;
