+++ src/test/java/info/mackiewicz/bankapp/system/security/password/PasswordServiceTest.java
@@ -1,5 +1,6 @@
 package info.mackiewicz.bankapp.system.security.password;
 
+import info.mackiewicz.bankapp.shared.service.PasswordService;
 import info.mackiewicz.bankapp.user.model.User;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
